requirements.txt
Flask==1.0.2
Pillow==8.1.0
